RT#22696: add (failing) test for inflation of 'back' references
The current git tree fails to fully inflate objects that are reached via
the 'back' field defined in an 'iset'. Take the following schema:

        Hat => {
             fields => {
                 string   => [ qw( colour ) ],
             },
        },
        NaturalPerson => {
             fields => {
                 string   => [ qw( firstName name ) ],
                 int      => [ qw( age ) ],
                 ref      => { partner => { null => 1 } },
                 iset    => {
                     hats => {
                         class => 'Hat',
                         back => 'owner',
                     },
                 },
             },
        },

If you $storage->select a 'Hat' you can reach the 'owner'. However the
owner's 'hats' array/Set::Object list is empty. The other fields are good.
This applies not just to hats but any other mapping that existed for the
class.

I know that this behaviour doesn't exist in 2.08 but don't know where
between 2.08 and git it broke.