RT#22696: use 'dummy' objects to collect items from @row
When we were creating temporary objects to call the importer methods,
these temporary object's destructors might have side effects, for
instance causing further demand paging.

So, we will now make a dummy object which gets passed to the
_row_to_object function, and make sure it gets blessed back to being a
dummy before we return.

This is an ugly work-around for the real problem, which is that the
importer protocol is too primitive.