Storage: handle another case where refaddrs are re-used
When objects fall out of scope, the weakref fires and removes the ref
from the 'objects' hash (mapping from OIDs to objects) but not the 'ids'
hash (mapping from refaddrs to OIDs).  If the refaddr is re-used by a
different object, but the first object has since been re-loaded at a
*different* address, then the check does not work.  The workaround will
likely affect programs that customized the 'set_id' schema property,
unfortunately.  It works by also storing a reference to the weakref
slot in the 'ids' hash, and checking that one instead.

Tracked down by Tomas Doran